scraper/backend/requirements.txt
aiofiles==23.1.0
algoliasearch==3.0.0
anyio==3.7.0
bcrypt==4.0.1
beanie==1.19.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.1.2
exceptiongroup==1.1.1
fastapi==0.99.0
fastapi_scheduler==0.0.12
h11==0.14.0
idna==3.4
inflect==6.0.4
Jinja2==3.1.2
joblib==1.2.0
lazy-model==0.0.5
motor==3.2.0
nltk==3.8.1
passlib==1.7.4
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pyfa-converter==1.0.4.1
pymongo==4.4.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.5
regex==2023.6.3
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.12.3
